requirements.txt
Faker==37.12.0
apache-airflow-providers-postgres
matplotlib
